ty Corthezz - structured in folders

+++ Injected/Objects/LocalPlayer.hpp
@@ -6,12 +6,12 @@ class LocalPlayer : public Player
 public:
 	LocalPlayer(int pointer) : Player(pointer) { }
 
-	long long PetGuid()
+	uint64 PetGuid()
 	{
 		if (addr && *(int*)(addr + 0x1858) & 1)
 		{
 			int petGuidPonter = *(int*)(addr + 0x1008);
-			return *(long long*)(petGuidPonter + 0x770);
+			return *(uint64*)(petGuidPonter + 0x770);
 		}
 		return 0;
 	}
